[jira] [HBASE-4867] A tool to merge configuration files
Summary: With our cluster configuration setup it would be good to have a tool
that would merge HBase configuration files so that files appearing later in the
list would override properties specified in earlier files. This way we could
merge an application-specific configuration file with a cluster-specific
configuration file (with the latter overriding the former) and produce a single
HBase configuration file to install on the cluster.

Test Plan: Run the tool on two configuration files (common and
cluster-specific). Use the resulting configuration on a dev cluster.

Reviewers: todd, Karthik, tedyu, stack, JIRA

Reviewed By: Karthik

CC: Karthik, mbautin, todd

Differential Revision: 537

git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1208832 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: a9ca67a2c57ae38e527a97779dc90cdd7d63f4f2