[jira] [HBASE-5382] Test that we always cache index and bloom blocks
Summary: This is a unit test that should have been part of HBASE-4683 but was
not committed. The original test was reviewed as part of
https://reviews.facebook.net/D807. Submitting unit test as a separate JIRA and
patch, and extending the scope of the test to also handle the case when block
cache is enabled for the column family.

Test Plan: Run unit tests

Reviewers: JIRA, jdcryans, lhofhansl, Liyin

Reviewed By: jdcryans

CC: jdcryans

Differential Revision: https://reviews.facebook.net/D1695

git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1242894 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: af0612f8865f479a0c633ea0ff9592667ad5f99c