[HBASE-5867] Improve Compaction Throttle Default
Summary:
We recently had a production issue where our compactions fell
behind because our compaction throttle was improperly tuned and
accidentally upgraded all compactions to the large pool. The default
from HBASE-3877 makes 1 bad assumption: the default number of flushed
files in a compaction. MinFilesToCompact should be taken into
consideration. As a default, it is less damaging for the large thread
to be slightly higher than it needs to be and only get timed-majors
versus having everything accidentally promoted.

Test Plan:  - mvn test

Reviewers: JIRA, Kannan, Liyin
Reviewed By: Kannan
CC: stack

Differential Revision: https://reviews.facebook.net/D2943

git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1338809 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 07962e24a57506226ebae18b5f5fd93ef99056f9