HBASE-13335 Use serverHasMoreResults context in SmallScanner and SmallReversedScanner.
Use the context passed back via ScanResponse that a RegionServer
fills in to denote whether or not more results existing in the
current Region. Add a simple factory to remove a static method
used across both SmallScanner and SmallReverseScanner. Add new
unit tests for both scanner classes to test scans with and
without the new context (as a quick backward-compatibility test).


Former-commit-id: ffdcc0095232caf7dc98448abb6ff57b0333e962