HBASE-14481 Deprecate HBase Wiki
We have an awesome community resource in our online book. It's maintained and looked after with
diligence. We also have an HBase section on the hadoop wiki that hasn't been updated since 2012.
Let's sift through the pages of the wiki, bring over any content that's still relevant and
not already present in the book, and kill the wiki.


Former-commit-id: c575e18ab82627b5f352caa10d964828fc5c0b6a