HBASE_16963 Fixup our hbase-personality post YETUS-322 flip verify_ result codes
This patch is about updating our hbase-personality for yetus.
Basic yetus methods flipped what was true and false but we did
not adjust our usage in accordance on an update. See
YETUS-322 flip verify_ result codes. I need this in place so I
can prove that there is nothing I have to do for this the parent
issue  HBASE-16955 after I commit HBASE-16952, the revamp of how
we do proto generation. After HBASE-16952 the protoc check will do the
right thing in all cases (pb2.5 vs pb3.1 and we run the protoc
in any module that has protos in it)


Former-commit-id: 339017f0f6f0970374055c2205b4ba42af2fabf7